+++ packaging/nix/flake.nix
@@ -14,7 +14,7 @@
       {
         packages.default = pkgs.rustPlatform.buildRustPackage {
           pname = "beachcomber";
-          version = "0.3.1";
+          version = "0.4.0";
 
           src = ../..;